refactor: 카테고리 상수 분리
