+++ src/pages/reservation/hooks/use-reservation.ts
@@ -9,7 +9,7 @@ import {
   extractLocationName,
 } from '@utils/extract-location';
 import { formatSelectedDateToSchedules } from '@utils/date-formatter';
-import { FOOD_TRUCK_CATEGORIES } from '@shared/constant/categories';
+import { FOOD_TRUCK_CATEGORIES } from '@shared/constant/food-truck-categories';
 import { filtersAtom, notFilteredAtom } from '@shared/store/filter-store';
 import { confirmedRegionsAtom } from '@shared/store/regions-store';
 import {
