+++ src/shared/constant/food-categories.ts
@@ -7,5 +7,3 @@ export const FOOD_CATEGORIES = {
   CAFE_DESSERT: '카페/디저트',
   ETC: '기타',
 } as const;
-
-export type FoodCategoriesKey = keyof typeof FOOD_CATEGORIES;
