+++ src/shared/constant/food-truck-categories.ts
@@ -1,0 +1,10 @@
+export const FOOD_TRUCK_CATEGORIES = {
+  ALL: '전체',
+  KOREAN: '한식',
+  CHINESE: '중식',
+  JAPANESE: '일식',
+  WESTERN: '양식',
+  SNACK: '분식',
+  CAFE_DESSERT: '카페/디저트',
+  ETC: '기타',
+} as const;
